+++ src/mcmc_loop.cpp
@@ -152,8 +152,8 @@ void mcmc_loop_multinomial(matrix<size_t> &Xorder_std, bool verbose, vector<vect
 
             // set nthread based on number of observations * mtry
             /*  double fake_p = (state->use_all) ? state->p : state->mtry;
-            if (state->n_y * fake_p < 1e5) { omp_set_num_threads( std::min(4, int(state->nthread)) ); }
-            else if (state->n_y * fake_p < 5e5 ) { omp_set_num_threads( std::min(6, int(state->nthread) ) ); }
+            if (state->n_y * fake_p < 1e5) { omp_set_num_threads( std::min(4, size_t(state->nthread)) ); }
+            else if (state->n_y * fake_p < 5e5 ) { omp_set_num_threads( std::min(6, size_t(state->nthread) ) ); }
             else {omp_set_num_threads(state->nthread); }*/
 
             // omp_set_max_active_levels(3);
